frontend/imgprocess4/requirements.txt
google-generativeai
Pillow
tenacity